refactor(perl-lsp-rs): split regex hover logic into SRP submodule (#0000)
